Scope keys to 14 bits, bound equ to its line, size addresses by mode
All three from Greptile's review, all reproduced before fixing.

A bare equ redefined whatever label came last. last_label_sz_ was never
cleared, so it survived the newline:

    foo: ld a,b
      equ 5          -> foo silently became 5
      ld a,foo       -> 3E 05

It is cleared at end of line now, and kept across a colon so "foo: equ 5"
still works.

Local scope keys collided every 128 global labels. The scope rode in one byte
masked to 0x7F, so scope 1 and scope 129 produced the same key and the second
routine's @l overwrote the first. Two bytes now, each biased so neither can be
zero -- the table stores keys as C strings -- giving 14 bits, and a file that
would exceed that is refused rather than silently aliased. The regression test
uses a forward reference on purpose: a backward one resolves before the far
scope can overwrite the key, so it would not have caught this.

An address operand ignored the mode. parse_number always wrote three bytes
while the symbolic path had moved to two in Z80 mode, so "ld hl,$1234" under
adl=0 emitted a stray fourth byte and shifted everything after it:

    zap       21 34 12 00 C9
    ez80asm   21 34 12 C9

+++ src/label_stack.c
@@ -18,7 +18,7 @@ void ls_destroy(label_stack* ls) {
 }
 
 bool ls_push(label_stack* ls, const char* label, int sz, int bpos,
-             int next, int line, fixup_kind kind, uint8_t scope, int anon) {
+             int next, int line, fixup_kind kind, uint16_t scope, int anon) {
     if (ls->pos_ == ls->sz_) {
         return false;
     }
